refactor: 상수 파일 통합
- constants/number로 통일
