+++ constants/number.ts
@@ -2,3 +2,5 @@ export const REDIS_EXPIRE_SECONDS = 86400;
 
 export const PROMISING_AVAILABLE_DATES_MAX = 12;
 export const PROMISING_USER_MAX = 10;
+
+export const UNKNOWN_USER_ID = 100000;
