--- constants/nums.ts
@@ -1,2 +1,0 @@
-
-export const UNKNOWN_USER_ID = 100000;
