+++ services/promise-user-service.ts
@@ -2,7 +2,7 @@ import { Op } from 'sequelize';
 import PromiseModel from '../models/promise';
 import PromiseUser from '../models/promise-user';
 import User from '../models/user';
-import {UNKNOWN_USER_ID} from '../constants/nums';
+import { UNKNOWN_USER_ID } from '../constants/number';
 
 class PromiseUserService {
   async findPromiseMembers(promises: PromiseModel[]) {
@@ -20,15 +20,15 @@ class PromiseUserService {
     }
     return promises;
   }
-  
-  async updateResignMember(userId:number){
+
+  async updateResignMember(userId: number) {
     const promiseUsers = await PromiseUser.findAll({
       where: {
         userId: userId
       }
     });
-    if(!promiseUsers) return;
-    await PromiseUser.update({userId:UNKNOWN_USER_ID}, {where: {userId:userId}})
+    if (!promiseUsers) return;
+    await PromiseUser.update({ userId: UNKNOWN_USER_ID }, { where: { userId: userId } });
   }
 }
